+++ src/lib/components/HyvorBar/BarUser.svelte
@@ -3,23 +3,27 @@
 	import ActionListItem from '../ActionList/ActionListItem.svelte';
 	import Dropdown from '../Dropdown/Dropdown.svelte';
 	import IconBoxArrowUpRight from '@hyvor/icons/IconBoxArrowUpRight';
-	import { barUser } from './bar.js';
 	import BarUserPreview from './BarUserPreview.svelte';
+	import BarUserPicture from './BarUserPicture.svelte';
 
 	interface Props {
 		instance: string;
+		logoutUrl?: string;
+		cloud: boolean;
 	}
 
-	let { instance }: Props = $props();
+	let { 
+		instance, 
+		logoutUrl = `${instance}/account/logout`,
+		cloud,
+	}: Props = $props();
 </script>
 
 <div class="wrap">
 	<Dropdown align="end" width={325}>
 		{#snippet trigger()}
 			<button class="user-wrap">
-				{#if $barUser}
-					<img class="user-picture" src={$barUser?.picture_url} alt={$barUser?.name} />
-				{/if}
+				<BarUserPicture />
 			</button>
 		{/snippet}
 
@@ -27,15 +31,19 @@
 			<ActionList>
 				<BarUserPreview />
 
-				<a href="{instance}/account" target="_blank">
-					<ActionListItem>
-						Manage Account
-						{#snippet end()}
-							<IconBoxArrowUpRight size={12} />
-						{/snippet}
-					</ActionListItem>
-				</a>
-				<a href="{instance}/account/logout">
+				{#if cloud}
+					<a href="{instance}/account" target="_blank">
+						<ActionListItem>
+							Manage Account
+							{#snippet end()}
+								<IconBoxArrowUpRight size={12} />
+							{/snippet}
+						</ActionListItem>
+					</a>
+				{/if}
+
+
+				<a href="{logoutUrl}">
 					<ActionListItem>Logout</ActionListItem>
 				</a>
 			</ActionList>
@@ -64,9 +72,5 @@
 	}
 	.user-wrap:hover {
 		box-shadow: 0 0 0 4px var(--input);
-	}
-	img {
-		width: 100%;
-		height: 100%;
 	}
 </style>