+++ src/routes/document-test/+page.svelte
@@ -5,7 +5,15 @@
 	import Footer from '$lib/marketing/Footer/Footer.svelte';
 </script>
 
-<HyvorBar product="core" />
+<HyvorBar product="core" cloud={false} authOverride={{
+	user: {
+		name: 'Test User',
+		// picture_url: 'https://picsum.photos/60/60',
+		picture_url: null,
+		email: 'test@hyvor.com'
+	},
+	logoutUrl: 'https://logout.com'
+}} />
 
 <Base>
 	<slot />